+++ .github/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2025-08-28
+
+### Added
+
+- HubSpot integrations.
+
+### Changed
+
+- Made the navigation menu more compact with Customers, Settings, and Shortcuts.
+
 ## [1.3.0] - 2024-05-27
 
 ### Added
@@ -30,7 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fixed a bug on the Customer table when adding new customers, the previously/newly added row would disappear due to a bug in the resetSort() function.
+- Fixed a bug on the Customer table when adding new customers, the previously/newly added row would disappear due to a bug in the `resetSort()` function.
 
 ## [1.1.3] - 2023-01-27
 
@@ -62,6 +72,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Release of Palm Tree.
 
+[2.0.0]: https://github.com/scottgriv/Palm-Tree/compare/v1.3.0...2.0.0
 [1.3.0]: https://github.com/scottgriv/Palm-Tree/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/scottgriv/Palm-Tree/compare/v1.1.4...v1.2.0
 [1.1.4]: https://github.com/scottgriv/Palm-Tree/compare/v1.1.3...v1.1.4